Clean up two sets of compilation warnings on Linux PPC.
git-svn-id: https://svn.parrot.org/parrot/trunk@10563 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ compilers/imcc/optimizer.c
@@ -589,7 +589,7 @@ eval_ins(Interp *interpreter, char *op, size_t ops, SymReg **r)
     for (i = 0; i < op_info->op_count - 1; i++) {
         switch (op_info->types[i]) {
             case PARROT_ARG_IC:
-                assert(i && ops == i);
+                assert(i && ops == (unsigned int)i);
                 /* set branch offset to zero */
                 eval[i + 1] = 0;
                 break;
